backend/server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
langchain
langchain-google-genai
langchain-chroma
langchain-community
pypdf
